2025_04_omeka_s_computer_vision/requirements.txt
pandas
numpy
dash
dash-bootstrap-components
plotly
datasets
pillow
transformers
sentence-transformers
einops
lancedb
requests
umap-learn
scikit-learn
hdbscan
git+https://github.com/gegedenice/omeka-s-api-client.git
python-dotenv